+++ Strongbox/Classes/Strongbox.swift
@@ -155,3 +155,52 @@ public class Strongbox {
     }
 
 }
+
+// Codable
+
+extension Strongbox {
+    public func encode<T: Encodable>(_ encodable: T?, key: String, encoder: JSONEncoder = .init(), accessibility: CFString = kSecAttrAccessibleWhenUnlocked) throws {
+        let sb = Strongbox()
+        guard
+            let encodable = encodable
+            else { sb.archive(nil, key: key, accessibility: accessibility) ; return }
+        let data = try encoder.encode(encodable)
+        sb.archive(data, key: key, accessibility: accessibility)
+    }
+    
+    public func decode<T: Decodable>(forKey key: String, decoder: JSONDecoder = .init()) throws -> T? {
+        let sb = Strongbox()
+        guard
+            let data = sb.unarchive(objectForKey: key) as? Data
+            else { return nil }
+        return try decoder.decode(T.self, from: data)
+    }
+}
+
+// Convenience 
+
+extension Strongbox {
+    
+    @discardableResult
+    public static func archive(_ object: Any?, key: String, accessibility: CFString = kSecAttrAccessibleWhenUnlocked) -> Bool {
+        let sb = Strongbox()
+        return sb.archive(object, key: key, accessibility: accessibility)
+    }
+    
+    public static func unarchive(objectForKey key: String) -> Any? {
+        let sb = Strongbox()
+        return sb.unarchive(objectForKey: key)
+    }
+}
+
+extension Strongbox {
+    public static func encode<T: Encodable>(_ encodable: T?, key: String, encoder: JSONEncoder = .init(), accessibility: CFString = kSecAttrAccessibleWhenUnlocked) throws {
+        let sb = Strongbox()
+        try sb.encode(encodable, key: key, encoder: encoder, accessibility: accessibility)
+    }
+    
+    public static func decode<T: Decodable>(forKey key: String, decoder: JSONDecoder = .init()) throws -> T? {
+        let sb = Strongbox()
+        return try sb.decode(forKey: key, decoder: decoder)
+    }
+}
